Disable default features on env_logger
The `env_logger` crate is used in tests and examples for logging output.
It has default features that rely on `is-terminal`, which relies on `rustix`.
`rustix` doesn't compile on the `esp*` targets, and thus using this project on such targets fails.

The simplest solution is to disable the default features.