mDNS: always set the "cache flush" bit
Always broadcast mDNS messages with the "cache flush" bit set for
TXT, SRV, A, and AAAA records. This is useful when the Matter device
reboots and gets assigned a different IP address, for example. In
such a case, hosts with an outdated cache may need to wait until the
original TTL expires (up to 60 seconds, per the TTL definitions on
mdns/builtin.rs).