Implement on off (#316)
This commit implements the On/Off cluster according to the Matter 1.3 spec. This PR also makes changes to the LevelControl cluster following changes in the coupling pattern, i.e. removing the LevelControlState.

The new usage patterns requires users to call an init method on the cluster handler after instantiation. This provides the user with the opportunity to declare the handler of a coupled cluster, validates the handler setup against the cluster configuration and executes required startup logic.