Updates the LevelControlHooks trait, separating the responsibility of setting the current_level attribute from setting the level of the device. Replaces Nullable with Option in the LevelControlHooks' StartUpOnOff assessors. Adds tests::LevelControlDeviceLogic. (#320)
This PR updates the `LevelControlHooks` trait, separating the responsibility of setting the `current_level` attribute from setting the level of the device. This is necessary to ensure that when the device is switched off via an OnOff command, the device is not switched on again when setting the CurrentLevel to the temporarily stored value.

From section 1.6.4.1.1
> Temporarily store CurrentLevel. Change CurrentLevel to the minimum level allowed for the device over the time period OnOffTransitionTime. If OnLevel is not defined, set the CurrentLevel to the stored level.

Additionally, this PR
- Replaces `Nullable` with `Option` in the `LevelControlHooks`' StartUpOnOff assessors, providing a more familiar Rust API.
- Adds a basic `LevelControlHooks` implementation (`tests::LevelControlDeviceLogic`) for use in tests and examples.