A guinea-pig binary for bloat checks (#332)
As per the documents inside the binary code, it is specially crafted so
that all memory except the stack is reported towards the `.bss` section.

For this to happen, all `rs-matter` types are allocated statically.
This includes the futures (whose size is anything but trivial), which
are moved/emplaced into `.bss` by the specific executor which is
utilized (`embassy-executor`).

TBDs:
- Check why the code size reported for the riscv32imac target is more
than twice as big as the thumb targets. Might just be due to the
compressed size of the thumb instructions (16bit) or there could be
something else going on for riscv32, as in linking to binaries we don't
need and use, but not removed by the LTO
- Fix the large "responder" future by splitting it into multiple tasks
in an embassy-executor task pool. This should reduce these 20KB down to
~ 10KB on x64 and even less for the other targets