More tests (#334)
This PR contains the following commits:

### 1st commit

Simply enables the 7 tests below. Did not require any changes to
`rs-matter`
* Test_AddNewFabricFromExistingFabric
* TestConfigVariables
* TestDelayCommands
* TestEqualities
* TestFabricRemovalWhileSubscribed
* TestSubscribe_AdministratorCommissioning
* TestSubscribe_OnOff

### 2nd commit

Enables the `TestDiscovery` test.
For this test, it was necessary to implement all optional mDNS payload:
- Device name
- Device type
- Device pairing hints (used to be hard-coded to "33")
- Device pairing information (used to be hard-coded to the empty string)

While I could've just disabled the tests for the optional payload in the
`chip_tests.pics` file, I thought it would be good to support this
payload in general, for end users. The payload is now configurable
through the `BasicInfo` structure.

### 3rd commit

This is a follow-up to the 2nd commit, which was not strictly necessary
but I did it anyway. It documents as well as does a cleanup on the
`pairing` module, which is all about computing the commissioning QR
code.

Some of the problems addressed with this commit:
- There were too many public freestanding functions that used to take a
lot of parameters (i.e. they were difficult to use). They are now all
moved into their appropriate structures (`Qr` or `QrPayload`)
- `QrTextPayload` renamed to just `QrPayload`. The purpose in life of
`QrPayload` is to compute the QR **text**
- A new type, `Qr` which is the final outcome of the QR computation; it
takes (a) text and encodes it into a QR; all utility functions to
consume the QR code as ASCII-art are now moved into this type and this
type is independent from `QrPayload` in that while it is expected that
it will be used to render precisely the text produced by `QrPayload`,
that might not be necessarily the case

### 4th commit

This commit introduces a real expiration of the Commissioning Window.
Previously, the timeout parameter was not really used and therefore, the
commissioning window stayed open until explicitly closed, or in case
commissioning did fail - forever.

Furthermore, I've tried to document and tighten up a bit the code in the
`case` and `pake` modules.
While I doubt this would be the latest cleanup in these modules (the
contain one of the oldest piece of code in the codebase), it is now in a
slightly better shape, in that a lot of freestanding functions are moved
to their appropriate structures, and some structures like `Spake2p` and
`CaseSession` are more private in that the SC layer doesn't need to know
about those. It only needs to know about `Pake` and `Case` which handle
the establishment of a PASE or CASE session, and `Spake2p` and
`CaseSession` are actually their internal implementation detail.


But the most important renaming change is that what used to be called
"PASE session" is now called "Commissioning Window" everywhere, because
really that's what it is.

The notion of a PASE session is anyway overloaded with two additional
meanings so **also** calling the Commissioning Window a "PASE session"
was wrong IMO:
- A PASE session might mean a `Session` which was created as a result of
a PASE/PAKE session negotiation/establishment. (Ditto - a CASE session
might mean a `Session` which was created as a result of a CASE session
negotiation/establishment.)
- A PASE session might also mean the process of **establishment** of a
`Session` using the PAKE negotiation process. Though I've taken care to
always talk there about a PASE session **establishment**, as in the
Matter Core Spec